Refactor the app to allow an unresolved default asset
This commit juggles some of the code around to allow for an unresolved
default asset. This is to allow the `configure` command to skip
validation before saving

However it means subsequent commands require the ability
to continue on from this state.

This is done by allowing the unresolved asset to be saved on the
Credentials object. The ability to resolve it has also been moved
on to the Credentials